requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
SQLAlchemy==2.0.34
python-dotenv==1.0.1
Werkzeug==3.0.4
email-validator==2.2.0
pytest==8.3.3
pytest-flask==1.3.0
gunicorn==23.0.0
